+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/laravel/sentinel/compare/v1.1.0...1.x)
 
+* Authorize local connections from Laravel Valet and Herd by [@Abil-official](https://github.com/Abil-official) in https://github.com/Abil-official/sentinel/pull/1
+
 ## [v1.1.0](https://github.com/laravel/sentinel/compare/v1.0.1...v1.1.0) - 2026-03-24
 
 * Fix typo in CI workflow exclusion matrix by [@JoshSalway](https://github.com/JoshSalway) in https://github.com/laravel/sentinel/pull/4